Get known question ids if they don't exist
